// apps/portal/src/api/generated/model/setVoteRulesetDtoQuorumElectorateBasis.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * HOA Manager API
 * The HOA Manager API description
 * OpenAPI spec version: 1.0
 */

export type SetVoteRulesetDtoQuorumElectorateBasis =
    (typeof SetVoteRulesetDtoQuorumElectorateBasis)[keyof typeof SetVoteRulesetDtoQuorumElectorateBasis];

export const SetVoteRulesetDtoQuorumElectorateBasis = {
    ALL_UNITS: "ALL_UNITS",
    ELIGIBLE_UNITS_ONLY: "ELIGIBLE_UNITS_ONLY",
} as const;