// apps/portal/src/api/generated/model/setVoteRulesetDtoWeightBasis.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * HOA Manager API
 * The HOA Manager API description
 * OpenAPI spec version: 1.0
 */

export type SetVoteRulesetDtoWeightBasis =
    (typeof SetVoteRulesetDtoWeightBasis)[keyof typeof SetVoteRulesetDtoWeightBasis];

export const SetVoteRulesetDtoWeightBasis = {
    UNIT_SHARE: "UNIT_SHARE",
    ONE_UNIT_ONE_VOTE: "ONE_UNIT_ONE_VOTE",
} as const;